feat: Refresh data on app visibility change
Adds an effect to the root App component that listens for the `visibilitychange` event. When the application returns to a visible state and a user is logged in, it forces a refresh of user data. This is intended to fix an issue where WebSockets might disconnect while the app is in the background, ensuring data is up-to-date upon foregrounding.